:hammer: Split blobPosts components
